Refactor wavefront module and isolate reference sphere calculation
This commit refactors the `wavefront.py` module into its own sub-package `optiland/wavefront/`. Each class has been moved to its own file to improve modularity and maintainability.

The reference sphere calculation has been isolated using a Strategy pattern. A new abstract base class, `ReferenceSphereCalculator`, has been introduced, along with a concrete implementation, `ChiefRayReferenceSphereCalculator`, which encapsulates the original calculation method. The `Wavefront` class now uses this strategy, allowing for future expansion with different calculation methods.

The public API is preserved, and a `DeprecationWarning` has been added to the old `wavefront.py` module to ensure backward compatibility.

All related tests have been updated to reflect these changes, and new tests have been added for the reference sphere calculation strategy. The code has been formatted and linted to adhere to project standards.
