+++ optiland/wavefront/__init__.py
@@ -1,0 +1,28 @@
+"""
+Wavefront subpackage for Optiland.
+
+This subpackage contains modules for wavefront analysis, including OPD (Optical
+Path Difference) calculations, Zernike polynomial fitting, and related data
+structures.
+
+Public API:
+- `Wavefront`: The main class for wavefront analysis.
+- `WavefrontData`: A data container for wavefront results.
+- `OPD`: A class for calculating and visualizing OPD maps.
+- `OPDFan`: A class for creating OPD fan plots.
+- `ZernikeOPD`: A class for Zernike decomposition of OPD data.
+"""
+
+from .opd import OPD
+from .opd_fan import OPDFan
+from .wavefront import Wavefront
+from .wavefront_data import WavefrontData
+from .zernike_opd import ZernikeOPD
+
+__all__ = [
+    "Wavefront",
+    "WavefrontData",
+    "OPD",
+    "OPDFan",
+    "ZernikeOPD",
+]
